Move action-group components to action-group folder
