+++ src/components/action-group/action-group/action-group.scss
@@ -1,4 +1,4 @@
-@import "../../common/_base";
+@import "../../../common/_base";
 
 @include box-sizing();
